back/requirements.txt
ansel==1.0.0
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
convertdate==2.4.0
cryptography==41.0.3
Django==4.2.3
djangorestframework==3.14.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
mod-wsgi==4.9.4
msgpack==1.0.5
mysql-connector==2.2.9
mysqlclient==2.2.0
numpy==1.25.1
pandas==2.0.3
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
PyMeeus==0.5.12
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pyzipper==0.3.6
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
